* release/1.1.0: (130 commits)
  updating poms for 1.1.0 release
  updating poms for 1.1.0 release
  changes to ensure backward compatibility with tomcat
  changes to ensure backward compatibility with tomcat
  upgrade product version
  changes to ensure backward compatibility with tomcat
  Update README.md
  Update README.md
  modify log pattern add MDC remove unnecesary debug traces in CompareDate
  change installation path
  fix jersey version in pom parent
  application management functionality
  delete managers not use add tests
  modify coveralls badge
  modify syntax
  add target path for rpm file to doc.
  adding tests
  change in the  verif
  work on the comments
  inclusion of a test for application instance dao and resource
  ...

+++ core/src/main/java/com/telefonica/euro_iaas/paasmanager/manager/impl/TierManagerImpl.java
@@ -30,7 +30,8 @@ import java.util.List;
 import java.util.Set;
 import java.util.StringTokenizer;
 
-import org.apache.log4j.Logger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import com.telefonica.euro_iaas.commons.dao.AlreadyExistsEntityException;
 import com.telefonica.euro_iaas.commons.dao.EntityNotFoundException;
@@ -71,7 +72,7 @@ public class TierManagerImpl implements TierManager {
 
     private SystemPropertiesProvider systemPropertiesProvider;
 
-    private static Logger log = Logger.getLogger(TierManagerImpl.class);
+    private static Logger log = LoggerFactory.getLogger(TierManagerImpl.class);
 
     /**
      * It add teh security groups related the products.
@@ -104,7 +105,7 @@ public class TierManagerImpl implements TierManager {
         log.debug("Create tier name " + tier.getName() + " image " + tier.getImage() + " flavour " + tier.getFlavour()
                 + " initial_number_instances " + tier.getInitialNumberInstances() + " maximum_number_instances "
                 + tier.getMaximumNumberInstances() + " minimum_number_instances " + tier.getMinimumNumberInstances()
-                + " floatingip " + tier.getFloatingip() + " keypair " + tier.getKeypair() + " icono " + tier.getIcono()
+                + " floatingIp " + tier.getFloatingip() + " keyPair " + tier.getKeypair() + " icon " + tier.getIcono()
                 + " product releases " + tier.getProductReleases() + "  vdc " + claudiaData.getVdc() + " networks "
                 + tier.getNetworks());
 
@@ -113,7 +114,7 @@ public class TierManagerImpl implements TierManager {
         } else {
 
             // check if exist product or need sync with SDC
-            existProductOrSyncWithSDC(tier);
+            existProductOrSyncWithSDC(claudiaData, tier);
 
             createSecurityGroups(claudiaData, tier);
 
@@ -124,13 +125,13 @@ public class TierManagerImpl implements TierManager {
         }
     }
 
-    private void existProductOrSyncWithSDC(Tier tier) throws InvalidEntityException {
+    private void existProductOrSyncWithSDC(ClaudiaData data, Tier tier) throws InvalidEntityException {
 
         if (tier.getProductReleases() != null && tier.getProductReleases().size() != 0) {
             for (ProductRelease prod : tier.getProductReleases()) {
                 try {
                     log.debug("Sync product release " + prod.getProduct() + "-" + prod.getVersion());
-                    prod = productReleaseManager.load(prod.getProduct() + "-" + prod.getVersion());
+                    prod = productReleaseManager.load(prod.getProduct() + "-" + prod.getVersion(), data);
                 } catch (Exception e2) {
                     String errorMessage = "The ProductRelease Object " + prod.getProduct() + "-" + prod.getVersion()
                             + " not exist in database";
@@ -209,9 +210,9 @@ public class TierManagerImpl implements TierManager {
         }
 
         for (Network network : networkToBeDeployed) {
-            if (networkManager.exists(network.getNetworkName(), network.getVdc())) {
+            if (networkManager.exists(network.getNetworkName(), network.getVdc(), tier.getRegion())) {
                 log.debug("the network " + network.getNetworkName() + " already exists");
-                network = networkManager.load(network.getNetworkName(), network.getVdc());
+                network = networkManager.load(network.getNetworkName(), network.getVdc(), tier.getRegion());
 
             } else {
                 network = networkManager.create(network);
@@ -237,7 +238,7 @@ public class TierManagerImpl implements TierManager {
 
             String mens = "It is not possible to delete the tier " + tier.getName() + " since it is not exist";
             log.error(mens);
-            throw new EntityNotFoundException(Tier.class, mens, tier);
+            throw new com.telefonica.euro_iaas.commons.dao.EntityNotFoundException(Tier.class, mens, tier);
         }
 
         if (tier.getSecurityGroup() != null && !tier.getVdc().isEmpty()) {
@@ -337,10 +338,6 @@ public class TierManagerImpl implements TierManager {
         List<Rule> rules = new ArrayList<Rule>();
         // 9990
         log.debug("Generate security rule " + 9990);
-
-        Rule rule = new Rule("TCP", "9990", "9990", "", systemPropertiesProvider.getProperty("sdcIp") + "/32");
-
-        rules.add(rule);
         Rule rule2 = new Rule("TCP", "22", "22", "", "0.0.0.0/0");
         rules.add(rule2);
         return rules;
@@ -351,14 +348,19 @@ public class TierManagerImpl implements TierManager {
 
         productRelease = productReleaseManager.loadWithMetadata(productRelease.getProduct() + "-"
                 + productRelease.getVersion());
-        Metadata openPortsAttribute = productRelease.getMetadata("open_ports");
+        getRules(productRelease, rules, "open_ports");
+        getRules(productRelease, rules, "open_ports_udp");
+
+    }
+
+    private void getRules(ProductRelease productRelease, List<Rule> rules, String pathrules) {
+        Metadata openPortsAttribute = productRelease.getMetadata(pathrules);
         if (openPortsAttribute != null) {
             log.debug("Adding product rule " + openPortsAttribute.getValue());
             StringTokenizer st = new StringTokenizer(openPortsAttribute.getValue());
             while (st.hasMoreTokens()) {
                 Rule rule = createRulePort(st.nextToken());
                 if (!rules.contains(rule)) {
-                    log.debug("New rule ");
                     rules.add(rule);
                 }
             }
@@ -481,12 +483,12 @@ public class TierManagerImpl implements TierManager {
                 throw new InvalidEntityException(errorMessage);
             }
 
-            if (productReleases != null && productReleases.size() != 0) {
+            if (productReleases.size() != 0) {
                 for (ProductRelease product : productReleases) {
 
                     try {
                         ProductRelease templateProduct = productReleaseManager.load(product.getProduct() + "-"
-                                + product.getVersion());
+                                + product.getVersion(), data);
                         log.debug("Adding product release " + templateProduct.getProduct() + "-"
                                 + templateProduct.getVersion() + " to tier " + templateProduct.getName());
 
@@ -510,7 +512,7 @@ public class TierManagerImpl implements TierManager {
         for (Network net : networskout) {
 
             try {
-                net = networkManager.load(net.getNetworkName(), net.getVdc());
+                net = networkManager.load(net.getNetworkName(), net.getVdc(), net.getRegion());
                 log.debug("Adding network " + net.getNetworkName() + "-" + " to tier " + tier.getName());
                 tier.addNetwork(net);
                 update(tier);
@@ -579,8 +581,8 @@ public class TierManagerImpl implements TierManager {
 
     }
 
-    public void updateTier(Tier tierold, Tier tiernew) throws InvalidEntityException, EntityNotFoundException,
-            AlreadyExistsEntityException {
+    public void updateTier(ClaudiaData data, Tier tierold, Tier tiernew) throws InvalidEntityException,
+            EntityNotFoundException, AlreadyExistsEntityException {
 
         tierold.setFlavour(tiernew.getFlavour());
         tierold.setFloatingip(tiernew.getFloatingip());
@@ -609,14 +611,16 @@ public class TierManagerImpl implements TierManager {
             try {
                 net = networkManager.create(net);
             } catch (AlreadyExistsEntityException e) {
-                net = networkManager.load(net.getNetworkName(), net.getVdc());
+                net = networkManager.load(net.getNetworkName(), net.getVdc(), net.getRegion());
             }
             tierold.addNetwork(net);
             update(tierold);
         }
 
         for (Network net : nets) {
-            networkManager.delete(net);
+            if (isAvailableToBeDeleted(net)) {
+                networkManager.delete(net);
+            }
         }
 
         tierold.setProductReleases(null);
@@ -627,8 +631,8 @@ public class TierManagerImpl implements TierManager {
 
         for (ProductRelease productRelease : tiernew.getProductReleases()) {
             try {
-                productRelease = productReleaseManager.load(productRelease.getProduct() + "-"
-                        + productRelease.getVersion());
+                productRelease = productReleaseManager.load(
+                        productRelease.getProduct() + "-" + productRelease.getVersion(), data);
             } catch (EntityNotFoundException e) {
                 log.error("The new software " + productRelease.getProduct() + "-" + productRelease.getVersion()
                         + " is not found");
